+++ src/main/java/com/fms/model/StorageDto.java
@@ -1,0 +1,13 @@
+package com.fms.model;
+
+import lombok.Data;
+import org.springframework.web.multipart.MultipartFile;
+
+@Data
+public class StorageDto {
+
+    private String tenant;
+    private byte[] content;
+    private MultipartFile multipartFile;
+
+}