+++ src/main/java/com/fms/service/storage/FilePerYearMonthStorageStrategyServiceService.java
@@ -1,0 +1,16 @@
+package com.fms.service.storage;
+
+import com.fms.model.StorageDto;
+import org.springframework.stereotype.Service;
+
+import java.util.Map;
+
+@Service
+public class FilePerYearMonthStorageStrategyServiceService implements StorageStrategyService {
+
+    @Override
+    public Map<String, String> store(StorageDto storageDto) {
+        return null;
+    }
+
+}