+++ src/main/java/com/fms/service/storage/StorageStrategyService.java
@@ -1,0 +1,11 @@
+package com.fms.service.storage;
+
+import com.fms.model.StorageDto;
+
+import java.util.Map;
+
+public interface StorageStrategyService {
+
+    Map<String, String> store(StorageDto storageDto);
+
+}